Added a fix for issue #5
By doing some string manipulation the lastfullbackup and lastdiffbackup
variables are compared against each other.  I could be all jacked up on
coffee right now but this seems like it should hold as the numerical
values of the dates will always make the most recent date greater.

Also commented out the TEMP folder name that AJ added because that
doesn't work unless that folder exists.  If we want to implement a
folder path option we should do that.